[FIX/#71] PersonDetailView 리펙토링
